% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arma_eigen}
\alias{arma_eigen}
\title{arma_eigen}
\usage{
arma_eigen(M)
}
\arguments{
\item{M}{input matrix = XtX}
}
\value{
v eigen vector
}
\description{
SVD calculation for X=UΣV, use XtX=VtΣV to calculate the right eigen vector
}
\author{
Xin Zhou \url{xinchoubiology@gmail.com}
}

