% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{average_linkage}
\alias{average_linkage}
\title{average_linkage}
\usage{
average_linkage(M)
}
\arguments{
\item{M}{correlation matrix}
}
\value{
Matrix
}
\description{
Parallel Computing measurement matrix for different linkage type : average linkages
}

