% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{averageLinkage}
\alias{averageLinkage}
\title{averageLinkage}
\usage{
averageLinkage(M)
}
\arguments{
\item{M}{correlation matrix}
}
\value{
Matrix
}
\description{
Parallel Computing measurement matrix for different linkage type : average linkages
}

