% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{beta_regress}
\alias{beta_regress}
\title{beta_regress}
\usage{
beta_regress(M, pv, svs, full = 0L)
}
\arguments{
\item{M}{m x n expression matrix; Each row represents probes and each col means samples}

\item{pv}{n x B design matrix; Each col means phenotype of interest, and if B >=2,
means we have B-1 permutations on covariates of interest}

\item{svs}{n x (p-1) design matrix; Each row represent samples and each col means parameters}

\item{full}{full output or coefficient only}
}
\description{
Linear regression on covariate of interest when comparing with other covariates
}
\author{
Xin Zhou
}

