% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{bootstrap.fit}
\alias{bootstrap.fit}
\title{bootstrap.fit}
\usage{
bootstrap.fit(dat.m = NULL, design = NULL, coef = 2, B = NULL)
}
\arguments{
\item{dat.m}{n x m matrix of methylation microarray}

\item{design}{design matrix for expression data matrix(data.m)}

\item{coef}{covariate of interest; Default = 2}

\item{B}{permutation number of covariates of interest}
}
\value{
list
        coefficient
        sigma
        df.residule
}
\description{
bootstrap wrapper
}

