% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bootstrap_regress}
\alias{bootstrap_regress}
\title{bootstrap_regress}
\usage{
bootstrap_regress(M, mod, modn, B)
}
\arguments{
\item{M}{m x n expression matrix; Each row represents probes and each col means samples}

\item{mod}{n x p design matrix;}

\item{modn}{n x (p-1) null design matrix - covariate of interest;}

\item{B}{n x B matrix; Bootstrap iterations index matrix}
}
\description{
Dimitris[Bootstrap hypothesis testing in regression models]
}
\details{
Bootstrap testing on regression model for covariate of interested
}
\author{
Xin Zhou
}

