% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bumps-class}
\alias{bumps-class}
\title{Class "bumps"
The class contains information of methyaltion differential related region}
\description{
Class "bumps"
The class contains information of methyaltion differential related region
}
\section{Slots}{

\describe{
\item{\code{table}}{statistic table of CpG regions}

\item{\code{null}}{list of null distribution (H0: no differential methylation)}

\item{\code{algorithm}}{"get.bumps" algorithm's argumrnts}
}}
\keyword{classes}

