% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clique_merge}
\alias{clique_merge}
\title{clique_merge}
\usage{
clique_merge(M)
}
\arguments{
\item{M}{Indicator Square Matrix}
}
\value{
index NumericVector
}
\description{
Parallel clique merge by indicator matrix by prefix-product
}

