% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{clusterMaker}
\alias{clusterMaker}
\title{clusterMaker}
\usage{
clusterMaker(chr, pos, maxGap = 500, names)
}
\arguments{
\item{chr}{Chromosome vector}

\item{pos}{position numeric vector}

\item{maxGap}{max gap length between 2 probes within a region}

\item{names}{probe names vector}
}
\value{
cluster data.frame of length m(m probes). Probes within same region have same region identities list.
}
\description{
clusterMaker function based on spatial distance
}
\details{
If you define the maxGap 500, so we can guarantee probes in different regions has distance >= maxGap.
         As the result, if we extend region upstream & downstream maxGap / 2, our extended regions have no overlap
}
\examples{
require(IlluminaHumanMethylation450kanno.ilmn12.hg19)
Location <- data.frame(IlluminaHumanMethylation450kanno.ilmn12.hg19@data$Locations)
Probes <- rownames(Location)[rownames(Location) \%in\% rownames(dat.m)]
BED <- cbind(Location[Probes, 1:2])
BED <- cbind(BED, dat.m[Probes, 4:3])
cluster <- clusterMaker(chr = BED$chr, pos = BED$pos, maxGap = 1000, names = rownames(BED))
}

