% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combp.R
\name{combine.pvalue}
\alias{combine.pvalue}
\title{combine.pvalue}
\usage{
combine.pvalue(dat.m = NULL, pvalues, cluster, chr, pos, names,
  method = c("spearman", "pearson", "kendall"),
  combine = c("stouffer_liptak", "zscore"), weight = NULL, cutoff = 0.1,
  cores = detectCores())
}
\arguments{
\item{dat.m}{n x m delta M|beta matrix for n CpG sites across 2*m paired different patient samples}

\item{pvalues}{pvalue column vector for each probe}

\item{cluster}{correlated cluster list calculated by \link{corrclusterMaker}
cluster list contains: {single probe cluster, multiple probes cluster}}

\item{chr}{chromosome vector}

\item{pos}{position vector}

\item{names}{probe names vector ; used in function \link{clusterMaker}}

\item{method}{correlation calculation method; c("spearman", "pearson", "kendall"); Default "spearman"}

\item{combine}{combine pvalue method; c("stouffer_liptak", "zscore")}

\item{weight}{NULL; weight provided as \code{1/sigma} of each probe. It means each probe's contribution
to defined regions}

\item{cutoff}{cutoff for dmr(differential methylated regions) detection's qvalue(fdr);
Default 0.1}

\item{cores}{core number for R backend combp algorithm}
}
\value{
combine-pval
}
\description{
combine-pvalue accepts a list of pvalues of probes, and correlation matrices
             are claculated by input data matrix. Then, it estimate combined pvalue for
             candidate regions defined in input region list
}
\details{
calculation combination pvalue of given regions(pvalues within region are correlated)
}

