% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{completeLinkage}
\alias{completeLinkage}
\title{completeLinkage}
\usage{
completeLinkage(M)
}
\arguments{
\item{M}{correlation matrix}
}
\value{
Matrix
}
\description{
Parallel Computing measurement matrix for different linkage type : complete linkage
}

