% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{corrClust}
\alias{corrClust}
\title{corrClust}
\usage{
corrClust(dat.m = NULL, cluster, cutoff = 0.8, type = c("vector", "list"),
  method = c("pearson", "spearman"))
}
\arguments{
\item{dat.m}{n x m delta M|beta matrix for n CpG sites across 2*m paired different patient samples}

\item{cluster}{vector of probes clusters By distance(maxGap) constraints; Default = 0.8}

\item{cutoff}{threshold to filter distance based cluster}

\item{type}{output type of cluster; c("vector", "list"); Default "list"}

\item{method}{correlation calculation method; c("pearson", "spearman", "kendall"); Default "spearman"}
}
\value{
list of 2 components
        cluster
        correlation matrix
}
\description{
By \code{clusterMaker}, we can generate probe cluster under constraint of maximum
             distance. Adding the correlation constraint, we will truncate exsited cluster and
             calculate their correlation matrix \code{sigma} out
}
\details{
correlated cluster maker. Build the CpG cluster by correlation

only clusters contain >= 2 probes can get a corrected p-value
}

