% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{corrclusterMaker}
\alias{corrclusterMaker}
\title{corrclusterMaker}
\usage{
corrclusterMaker(dat.m = NULL, chr, pos, cluster = NULL, maxGap = 500,
  names, cutoff = 0.8, cores = 2, method = c("pearson", "spearman",
  "kendall"), merge = c("single", "complete", "average"))
}
\arguments{
\item{dat.m}{n x m delta M|beta matrix for n CpG sites across 2*m paired different patient samples}

\item{chr}{Chromosome vector}

\item{pos}{position numeric vector}

\item{cluster}{vector of probes clusters By distance(maxGap) constraints}

\item{maxGap}{max gap length between 2 probes within a region ; Default = 500
also used in function \link{clusterMaker}}

\item{names}{probe names vector ; used in function \link{clusterMaker}}

\item{cutoff}{threshold to filter distance based cluster; Default = 0.8}

\item{cores}{number of thread used by \link{corrclusterMaker}}

\item{method}{correlation calculation method; c("spearman", "pearson", "kendall"); Default "spearman"}

\item{merge}{how to merge two sub-clusters; c("single", "complete", "average")}

\item{pos}{CpG position with their probes id}
}
\value{
list of clusterID vector
}
\description{
By \code{clusterMaker}, we can generate probe cluster under constraint of maximum
             distance. Adding the correlation constraint, we will truncate exsited cluster and
             calculate their correlation matrix \code{sigma} out
}
\details{
correlated cluster maker. Build the CpG cluster by correlation

only clusters contain >= 2 probes can get a corrected p-value
}

