% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{Dbpmerge}
\alias{Dbpmerge}
\title{Dbpmerge}
\usage{
Dbpmerge(c.mat = NULL, merge = c("single", "complete", "average"),
  cutoff = 0.8)
}
\arguments{
\item{c.mat}{list of correlation probes matrices}

\item{merge}{merge method; c("single", "complete", "average")}

\item{cutoff}{correlation >= cutoff can be merge into a sub-regions}
}
\value{
list of all sub-clusters
}
\description{
sub-regions merge and split
}

