% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{dendro.data}
\alias{dendro.data}
\alias{dendro.data.default}
\alias{dendro.data.hclust}
\title{dendro.data}
\usage{
dendro.data(model, ...)

\method{dendro.data}{default}(model, ...)

\method{dendro.data}{hclust}(model, ...)
}
\arguments{
\item{model}{object of \code{\link{stats}{hclust}}}

\item{...}{ignored}

\item{model}{model object}

\item{model}{model object; hc}
}
\value{
ggplot of clustering hierachical tree
}
\description{
extract cluster data frame from a model into a list of data frame

default dendro data convert
dendro.data method for \code{default} object

dendro data convert for \code{\link{stats}{hclust}} object
}

