% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_kinds}
\alias{distance_kinds}
\title{distance_kinds}
\usage{
distance_kinds()
}
\value{
distance_type character vector for distance definition
}
\description{
Defined distance kinds for beta value matrix calculation
}

