% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{distributeRef}
\alias{distributeRef}
\title{distributeRef}
\usage{
distributeRef(edat = NULL, size = 10, by = "R", n = 1000, mcores = 2,
  ...)
}
\arguments{
\item{edat}{data matrix to sampled}

\item{size}{bootstrap sampling size}

\item{by}{data will independent by row('R')[Default] or by column}

\item{n}{restrict sampled data matrix size}

\item{mcores}{parallel threads to be used}
}
\value{
list of NULL dendrogram objects
}
\description{
To calculate the optimal cluster number of our data, we need a reference distribution
generation for NULL_DENDROGRAM
}
\author{
Xin Zhou \url{xxz220@miami.edu}
}

