% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clustering.R
\name{Eigenbeta}
\alias{Eigenbeta}
\title{Eigenbeta}
\usage{
Eigenbeta(mset = NULL, cluster = NULL, nPC = 1, verbose = TRUE,
  softpower = 6, corr = c("pearson", "spearman", "kendall"),
  align = c("along average", ""))
}
\arguments{
\item{mset}{normalized methylation matrix of 485512 CpG sites}

\item{cluster}{List; clustered neighboured CpG sites}

\item{nPC}{number of principle components}

\item{verbose}{Optional; TRUE(default), whether to print processing or not}

\item{softpower}{6; For hub beta vector generation}

\item{corr}{Correlation method; Optional; "pearson" default; "spearman" and "kendall"}

\item{align}{Use the average beta(Methylation) value to control the orientation of eigenunit
align =  "along average"(default, means that eigenunit should be aligned with average beta value)
align =  "" (means letting the eigenunit vector as themselves)}
}
\value{
list
        eigenBeta eigen beta vectors' matrix for pre-clustered prototype units
        averageBeta average beta vectors for each clustered CpG unit
}
\description{
CpG sites are clustered by bump-hunter or combp. So the correlation(similarity) network
             will be represented by eigen unit(which is the 1st PC)
}
\details{
get eigen-gene from pre-clustered prototype [Methods: BMC2007_Steve_Horvath]
and the second-layer network analysis will be based on [eigengene network]
}
\author{
Xin Zhou \url{xxz220@miami.edu}
}

