% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iSVA.R
\name{estDimRMT}
\alias{estDimRMT}
\title{estDimRMT}
\usage{
estDimRMT(mat, plot = TRUE)
}
\arguments{
\item{mat}{n x m M|beta matrix for n CpG sites across m different patient samples.}

\item{plot}{Optional bool value, plot eigenvalue distribution or not. Default = TRUE}
}
\value{
cor Cross correlation matrix across different samples

dim Estimate dimension of data

empeigen Empirical distribution of eigenvalue

theigen Theoretical distribution of eigenvalue

evalue Eigen values of correlation matrix

ggplot Options ggplot2 object
}
\description{
estimate the dimension of Methylation status methylation by RMT
}
\author{
Xin Zhou \url{xinchoubiology@gmail.com}
}

