% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{get.bumps}
\alias{get.bumps}
\title{get.bumps method}
\usage{
get.bumps(object, ...)
}
\arguments{
\item{object}{methylation expression matrix object}

\item{...}{additional parameters}
}
\value{
bumps object
}
\description{
get.bumps method
}

