% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get450K.R
\name{get450Ksheet}
\alias{get450Ksheet}
\title{get450Ksheet}
\usage{
get450Ksheet(save = ".", sdrf = NULL, verbose = FALSE)
}
\arguments{
\item{save}{sample sheet file save directory}

\item{sdrf}{sdrf directionary}
}
\value{
sample table
}
\description{
Rcppsva is compatible with minif package, as the result \code{get450Ksheet} is a function
             for samplesheet.csv preparation for user's methylation 450K microarray raw data
}
\details{
parse sample sheet and minfi compatible
}

