% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get450K.R
\name{get450Kurl}
\alias{get450Kurl}
\title{get450Kurl}
\usage{
get450Kurl(type = NULL, save = ".", verbose = FALSE, ...)
}
\arguments{
\item{type}{cancer type of TCGA 450K database}

\item{save}{save directionary}

\item{verbose}{FALSE(default)}
}
\value{
absolute path of tcga data archive for specific cancer type
}
\description{
\code{get450Kurl} function is used for specific TCGA archive meta-database file(.sdrf)
             downloading
}
\details{
get 450K url
}

