% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get450K.R
\name{get450Kzip}
\alias{get450Kzip}
\title{get450Kzip}
\usage{
get450Kzip(pdata = NULL)
}
\arguments{
\item{pdata}{Level 1 methylation 450K data matrix for pair extraction}
}
\value{
matrix for paired methylation data
}
\description{
when the ComBat analysis has been completed, we only needed to get the paired
methylation data for future research
}

