% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get450K.R
\name{getTCGAssemble}
\alias{getTCGAssemble}
\title{getTCGAssemble}
\usage{
getTCGAssemble(...)
}
\arguments{
\item{...}{}
}
\value{
string TCGA traverse result path
}
\description{
TCGA Assembler is a two-R-script module for TCGA data indexing. \code{getTCGAssemble}
             is used for TCGA Assembler environment preparation and directionary file retrieval
}
\details{
get TCGA-Assembler path
}

