% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{\%--\%}
\alias{\%--\%}
\title{Broadcast subtraction}
\usage{
x %--% y
}
\arguments{
\item{x}{numeric or complex vector}

\item{y}{numeric or complex vector}
}
\description{
broadcast operation on matrix
}

