% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{hclust_from_data}
\alias{hclust_from_data}
\title{hclust_from_data}
\usage{
hclust_from_data(data, link, dist, minkowski)
}
\arguments{
\item{data}{beta value matrix}

\item{link}{character for linkage type[default "ward" out of
c("ward", "average", "complete", "single")]}

\item{dist}{character for distance type[default "euclidean" out of
c("euclidean", "manhanttan")]}

\item{minkowski}{soft_power of minkowski distance}
}
\value{
hclust object for our cluster
}
\description{
execute hierachical clustering by beta value data directly
}

