% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{int.eprior}
\alias{int.eprior}
\title{Monte Carlo integration functions}
\usage{
int.eprior(sdat, g.hat, d.hat)
}
\arguments{
\item{g.hat}{estimated gamma values}

\item{d.hat}{estimated delta values}

\item{sdata}{standard data matrix}
}
\value{
adjust data frame
        g.star
        d.star
}
\description{
Monte Carlo integration functions
}

