% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isva.R
\name{isvaFn}
\alias{isvaFn}
\title{isvaFn}
\usage{
isvaFn(dat.m = NULL, design = NULL, type = c("M", "beta"),
  qcutoff = 0.05, verbose = FALSE)
}
\arguments{
\item{dat.m}{Data matrix whose rows represent different labeling features and columns stand for the different samples}

\item{design}{design matrix for expression/methylation microarray}

\item{qcutoff}{qvalue's cutoff, control the FDR(false discovery rate)}
}
\value{
n.isv number of ISVs

isv matrix of ISV
}
\description{
isvaFunction
}
\author{
Xin Zhou \url{xinchoubiology@gmail.com}
}

