% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{linkage_kinds}
\alias{linkage_kinds}
\title{linkage_kinds}
\usage{
linkage_kinds()
}
\value{
linkage_type character vector for linkage definition
}
\description{
Defined linkage kinds in hierachical cluster
}

