% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{List2Index}
\alias{List2Index}
\title{List2Index}
\usage{
List2Index(cluster, chr, pos, names)
}
\arguments{
\item{cluster}{}

\item{chr}{}

\item{pos}{}

\item{names}{}
}
\value{
clusterIndex
}
\description{
convert cluster list to Index with increased number
}

