% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_rank}
\alias{matrix_rank}
\title{matrix_rank}
\usage{
matrix_rank(M, by)
}
\arguments{
\item{M}{matrix for calculation rank}

\item{by}{row(1, default); col(2)}
}
\value{
ranked_matrix matrix represented by each (row/col) order
}
\description{
Give an order of Matrix by row | column
Question : how to deal with the vector contain multiple tie values
}

