% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{memberDend}
\alias{memberDend}
\title{memberDend}
\usage{
memberDend(subtree)
}
\arguments{
\item{subtree}{}
}
\value{
number in tree object x
}
\description{
return member number in subtree x
}

