% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{memberLimit}
\alias{memberLimit}
\title{memberLimit}
\usage{
memberLimit(x1, x2, subtree)
}
\arguments{
\item{x1}{leftmost node}

\item{x2}{rightmost node}

\item{subtree}{subtree of dendrogram}
}
\value{
list
        x
        limit c(tree1, tree2, tree3)
}
\description{
assign position of fork and two limit of tree to members
}

