% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{mlm.fit}
\alias{mlm.fit}
\title{mlm.fit}
\usage{
mlm.fit(dat.m = NULL, design = NULL, coef = 2, B = NULL, full = FALSE,
  mcore = 4)
}
\arguments{
\item{dat.m}{n x m matrix of methylation microarray}

\item{design}{design matrix for expression data matrix(data.m)}

\item{coef}{covariate of interest; Default = 2}

\item{B}{permutation number of covariates of interest}

\item{full}{return full regression object; Optional FALSE}
}
\value{
list
        coefficient
        stdev_unscale
        sigma
        df.residule
}
\description{
estimate coefficient of interest variables with known Svs, \link{mlm.tstat}
}
\examples{
sd <- 0.3 * sqrt(4/rchisq(100, df = 4))
y  <- matrix(rnorm(100*6, sd = sd), 100, 6)  # each row of data is generate by sd[i] ~ invchisq
rownames(y) <- paste("cg", 1:100)
# introduce 2 cgs which are DMPs
y[1:2, 4:6] <- y[1:2, 4:6] + 2 # have significant differential when we introduce the poi(cancer - normal)
pheno <- factor(c(0,0,0,1,1,1))
levels(pheno) <- c("normal", "cancer")
design <- model.matrix(~ pheno)
fit <- mlm.fit(y, design)
sig.tab <- mlm.tstat(fit)
library(doMC)
registerDoMC(2)
}

