% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{mlm.tstat}
\alias{mlm.tstat}
\title{mlm.tstat}
\usage{
mlm.tstat(fit)
}
\arguments{
\item{fit}{object of mlm.fit; contains coefficient, sigma}
}
\value{
list of 2 components
        post.sigma
        df.total
        t score
}
\description{
moderest lm T statistic
}

