% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clustering.R
\name{moduleSearch}
\alias{moduleSearch}
\title{moduleSearch}
\usage{
moduleSearch(beta.expr = NULL, cor.type = c("pearson", "spearman"),
  sim.type = c("signed", "unsigned"), method = c("average", "ward",
  "single", "complete"), deepSplit = 1, cutHeight = 0.9,
  minClusterSize = min(10, ncol(beta.expr)/2), pamStage = TRUE,
  pamRespectsDendro = TRUE, verbose = 2, ...)
}
\arguments{
\item{beta.expr}{pre-clustered genome region for co-regulation methylation status}

\item{cor.type}{correlation type for dissimilarity calculation; "pearson"(Default) and "spearman"}

\item{sim.type}{similarity type for hierachical clustering; "signed"(Default) and "unsigned"}

\item{method}{hierachical clustering method; c("average", "ward", "single", "complete")}

\item{deepSplit}{[0,1,2,3,4]. Parameter in cutreeDynamic to control the sensitivity of module detection}

\item{cutHeight}{h_{max} in DynamicTreeCut method}

\item{minClusterSize}{c_{min} minimum size of each cluster}

\item{pamStage}{Only used for method "hybrid" dynamicTreecut; If TRUE, the (PAM-like)stage will be
performed}

\item{pamRespectsDendro}{Logical}

\item{usebigmemory}{use bigmemory or not}
}
\value{
A list with following component:
        clusters  a vector of cluster labels for all CpG regions
        dendrogram dendrogram of Rclusterpp.hclust
}
\description{
cutreeDynamic is a dynamic clustering method borrowed from package dynamicTreeCut,
             and in this method, we will clustered our pre-clustered beta matrix and calculate
             out their co-regulation methylation regions
}
\details{
based on WGCNA and Dynamic Tree Cut, clustering the DMRs
Dynamic Tree cut method to determine co-methylation modules of different regions
}
\author{
Xin Zhou \url{xxz220@miami.edu}
}

