% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optimization-methods.R
\name{optimModule}
\alias{optimModule}
\title{optimModule}
\usage{
optimModule(gstat = NULL, dendro = NULL, plot = TRUE, verbose = TRUE)
}
\arguments{
\item{gstat}{gap statistic data}

\item{dendro}{dendrogram object of actual clustering}

\item{plot}{Optional, plot or not}
}
\value{
module object
}
\description{
optimal cluster number retrieve
}

