% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.corr}
\alias{plot.corr}
\title{plot correlation matrix by ggplot}
\usage{
\method{plot}{corr}(x, y = NULL, cutoff = 0.8, ...)
}
\arguments{
\item{x}{n x n correlation matrix}

\item{y}{NULL}

\item{cutoff}{correlation cutoff; Default = 0.8}
}
\value{
ggplot object
}
\description{
plot correlation matrix by ggplot
}
\examples{
corrmat <- list(cor = matrix(c(-0.9,-0.6,0.3,1), nrow = 2))
colnames(corrmat$cor) <- rownames(corrmat$cor) <- c("X1","X2")
class(corrmat) <- "corr"
plot(corrmat)
}

