% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{preProcess450K}
\alias{preProcess450K}
\title{preProcess450K}
\usage{
preProcess450K(tcga.dir = ".", cancer = "BRCA", methylate = c("Beta",
  "M"), verbose = TRUE)
}
\arguments{
\item{tcga.dir}{directionary of TCGA cancer level-1 data}

\item{cancer}{[ca]ncer [type] for TCGA retrieval}

\item{methylate}{[me]thylate data value type; 'Beta'(default) and 'M'}

\item{verbose}{TRUE}
}
\value{
list
        mset Methylated value matrix after ComBat process
        pd   pdata for next analysis
}
\description{
\code{process450K} is a function used for our methylation 450K raw data preprocessing.
}
\details{
utility R function; preProcess450K
}

