% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{preProcessDMR}
\alias{preProcessDMR}
\title{preProcessDMR}
\usage{
preProcessDMR(data = NULL, pdata = NULL, svam = c("isva", "sva"),
  lmfit = c("NULL", "limma"), sampling = c("permutation", "bootstrap"),
  B = 30, verbose = TRUE)
}
\arguments{
\item{data}{normalized methylation microarray value matrix (Beta, M)}

\item{pdata}{phenotype data matrix}

\item{svam}{surrogate variable analysis method type}

\item{lmfit}{linear regression analysis}

\item{sampling}{sampling coefficient type; 'permutation'(default) and 'bootstrap'}

\item{B}{sampling method's round; 30(default)}

\item{verbose}{TRUE}
}
\value{
list
        fit mset.fit matrix
        bg  mset.bg list
            coef each round's coeffcient
            sigma
            df.residuals
}
\description{
ultility R differential methylation positions 450K for M value data archive
}

