% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.bed}
\alias{print.bed}
\title{print.bed}
\usage{
print.bed(dat.m = NULL, bed = NULL)
}
\arguments{
\item{dat.m}{m x n matrix. Statitical test p-value and each row's name is its probes name}

\item{bed}{BED file name}
}
\value{
BED data.frame
}
\description{
output 450k microarray data frame with ordered positions to BED file
}

