% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{print.bed}
\alias{print.bed}
\title{output 450k microarray data frame with ordered positions to BED file}
\usage{
\method{print}{bed}(x, ...)
}
\arguments{
\item{x}{m x n matrix. Statitical test p-value and each row's name is its probes name}

\item{...}{bed BED file name}
}
\value{
BED data.frame
}
\description{
output 450k microarray data frame with ordered positions to BED file
}

