% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{protoGen}
\alias{protoGen}
\title{protoGen}
\usage{
protoGen(dat.m = NULL, prototype = NULL, method = c("pearson",
  "spearman"), mcores = 4)
}
\arguments{
\item{dat.m}{data matrix of methylation data}

\item{prototype}{Index list of each pre-clustered CpG sites unit}

\item{method}{Optional; pearson(Default) and spearman}

\item{mcores}{multiple threads used}
}
\value{
bmu matrix of clustered center for each pre-defined cluster
}
\description{
considering that clusters are determined by different correlation definition
             pearson correlation  : defined by mean
             spearman correlation : defined by median of rank
}
\details{
get pre-clustering matrix, here we follow SOM's definition of best matched unit based on
genome sequence constraint (Prototype Generation)
In pre-processing steps, we have generate the pre-clustered CpG sites by their spatial constraint,
therefore, the original data matrix should be modified for future processing
}
\author{
Xin Zhou \url{xxz220@miami.edu}
}

