% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{pvalue}
\alias{pvalue}
\title{pvalue}
\usage{
pvalue(dat.m = NULL, mod = NULL, mod0 = NULL, ...)
}
\arguments{
\item{dat.m}{data matrix for expression/methylation microarray}

\item{mod}{the model being used to fit the data}

\item{mod0}{the null hypothesis model being compared when fitting the data}

\item{verbose}{Optional; Default FALSE}
}
\value{
p vector of F-statistic p-values for each row of dat
}
\description{
Quick calculate f statistic p-values for mod and mod0
}
\author{
Xin Zhou
}

