% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rankm}
\alias{rankm}
\title{rankm}
\usage{
rankm(M, byrow)
}
\arguments{
\item{M}{matrix for calculation rank}

\item{byrow}{TRUE(FALSE; By column)}
}
\value{
ranked_matrix matrix represented by each (row/col) order
}
\description{
Give an order of Matrix by row | column
}

