% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{regionSeeker}
\alias{regionSeeker}
\title{regionSeeker}
\usage{
regionSeeker(beta, chr, pos, cluster = NULL, maxGap = 500, names,
  cutoff = c(quantile(abs(beta), 0.9), quantile(abs(beta), 0.1)),
  permbeta = NULL, drop = FALSE, verbose = FALSE)
}
\arguments{
\item{beta}{vector of different probes}

\item{chr}{Chromosome vector}

\item{pos}{position numeric vector}

\item{cluster}{list clusters defined by arguments or \link{clusterMaker}}

\item{maxGap}{max gap length between 2 probes within a region ;
used in function \link{clusterMaker}}

\item{names}{probe names vector ; used in function \link{clusterMaker}}

\item{cutoff}{cutoff threshold for bump selection. only segements within cluster satisfy cutoff
are returned as predicted region c(sig-cutoff, null-cutoff)}

\item{permbeta}{Null hypothesis beta distribution
[added permbeta(H0 distribution) to the end columns of region table from \link{regionSeeker}]}

\item{drop}{FALSE; create discriminate table (significant | null) or not}

\item{verbose}{FALSE}
}
\value{
Table of predict regions
}
\description{
bump hunting algorithm one clusters predefined by distance/correlation constraints
}
\details{
If an arbitary threshold is defined, regionSeeker will return a table (within / without)
         the threshold. In bump hunting algorithms, these contiguous probes mean bumps.
}

