% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{segmentsCluster}
\alias{segmentsCluster}
\title{segmentsCluster}
\usage{
segmentsCluster(cluster, beta, chr, pos, permbeta)
}
\arguments{
\item{cluster}{vector of probes clusters}

\item{beta}{beta value for probes}

\item{chr}{chromosome vector}

\item{pos}{position vector}

\item{permbeta}{null hypothesis distribution}
}
\value{
data.table of all clusters
}
\description{
segmentsCluster clustering correlated probes
}

