% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{segmentsMaker}
\alias{segmentsMaker}
\title{segmentsMaker}
\usage{
segmentsMaker(cluster, beta, cutoff, chr, pos, permbeta)
}
\arguments{
\item{cluster}{vector of probes clusters}

\item{beta}{beta value for probes}

\item{cutoff}{cutoff value for positive & negative value;
<= cutoff[1] : hypo
>= cutoff[4] : hyper
(cutoff[2], cutoff[3]) : null}

\item{chr}{chromosome vector}

\item{pos}{position vector}

\item{permbeta}{null hypothesis distribution}
}
\value{
cluster list and each cluster contains 3 list
        `hyper`| `+`
        `hypo` | `-`
        `null` | `0`
}
\description{
segment vectors into positive, null and negative
}
\details{
bumphunting kernel; and we can use comb-p method in segmentMaker.
}

