% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{single_linkage}
\alias{single_linkage}
\title{single_linkage}
\usage{
single_linkage(M)
}
\arguments{
\item{M}{correlation matrix}
}
\value{
Matrix
}
\description{
Parallel Computing measurement matrix for different linkage type : single linkage
}

