% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/smooth.R
\name{smoother}
\alias{smoother}
\title{smoother}
\usage{
smoother(beta, pos, names, cluster, weight, method = c("weightedLowess",
  "loess", "locfit"), mcores = detectCores())
}
\arguments{
\item{beta}{beta value for probes}

\item{pos}{position vector}

\item{names}{probe names vector ; used in function \link{clusterMaker} and \link{regionSeeker}}

\item{cluster}{cluster object for local regression}

\item{weight}{weight for local regression}

\item{method}{c("weightedLowess", "loess", "locfit")}

\item{mcores}{thread used in function}
}
\description{
robust local weighted regression, fitting a smooth nonparamatic regression
and smoothing methylation measurement.
}

