% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combp.R
\name{stouffer_liptak.combp}
\alias{stouffer_liptak.combp}
\title{stouffer_liptak.combp}
\usage{
stouffer_liptak.combp(pvalues, sigma, weight = NULL)
}
\arguments{
\item{pvalues}{vector of pvalues}

\item{sigma}{covariance of pvalue}

\item{weight}{use weight for pvalue combination; Default NULL}
}
\value{
Cp (Combined pvalue)
}
\description{
2 combp functions are borrowed from Combp value function of brentp
}
\details{
Calculate combined p-value by stouffer_liptak method

if \code{sigma} is not positive definitive, off-diag elements x 0.99
}

