% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isva.R
\name{svaReg}
\alias{svaReg}
\title{svaReg}
\usage{
svaReg(dat.m = NULL, design = NULL, sv.m = NULL, qvalue0 = 0.1,
  backend = c("NULL", "limma"), verbose = FALSE)
}
\arguments{
\item{dat.m}{n x m M|beta matrix for n CpG sites across m different patient samples.}

\item{design}{phenotype of interested; m-length vector reresent patient samples' phenotype.
Optional: [combine with sample pair information] and mandatory column 2 is covariate
of interest}

\item{sv.m}{surrogate variables matrix calculate from \link{isvaFn} other sva methods.}

\item{qvalue0}{false discovery rate's threshold; Default = 0.1}

\item{backend}{backend regression packages; Default = "NULL", switch to limma for moderate statistic}

\item{verbose}{Optional; Default FALSE}
}
\value{
res data.frame
}
\description{
when we search all of the factors which means surrogate variables
             update our model.matrix and calculate moderate test
}
\details{
svaReg Do regression with selected surrogate variables
}
\author{
Xin Zhou \url{xinchoubiology@gmail.com}
}

