% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{wilcox.fit}
\alias{wilcox.fit}
\title{wilcox.fit}
\usage{
wilcox.fit(dat.m = NULL, alternative = c("two.sided", "greater", "less"),
  qvalue0 = 0.1, mcore = 4)
}
\arguments{
\item{dat.m}{n x m delta M|beta matrix for n CpG sites across 2*m paired different patient samples}

\item{alternative}{a character string specifying the alternative hypothesis;
c("two.sided", "greater", "less") and "two.sided" is default}

\item{qvalue0}{false discovery rate's threshold; Default = 0.1}

\item{mcore}{Cpu cores can be used in test}
}
\value{
res list
            siggene data.frame
            null    data.frame
}
\description{
calculate wolcoxon signed rank sum test for each CpG probes of microarray;
             support multiple core parallism calculation.
}
\details{
If our data is paired, wilcoxon rank test can be used to test the significnat of sites
}
\author{
Xin Zhou
}

