% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combp.R
\name{zscore.combp}
\alias{zscore.combp}
\title{zscore.combp}
\usage{
zscore.combp(pvalues, sigma, weight = NULL)
}
\arguments{
\item{pvalues}{vector of pvalues}

\item{sigma}{covariance of pvalue}

\item{weight}{use weight for pvalue combination; Default NULL}
}
\value{
Cp (Combined pvalue)
}
\description{
Calculate the combined pvalue by combined Z score method
}
\details{
if \code{sigma} is not positive definitive, off-diag elements x 0.99
}

