% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bumphunter.R
\name{bumphuntingEngine}
\alias{bumphuntingEngine}
\title{bumphuntingEngine}
\usage{
bumphuntingEngine(dat.m = NULL, design, sv.m = NULL, chr, pos,
  cluster = NULL, coef = 2, cutoff = NULL, pvalue = 0.01,
  maxGap = 500, minDist = 500, nullMethod = c("permutation", "bootstrap"),
  robust = FALSE, smooth = FALSE, smoother = bumphunter::loessByCluster,
  B = 10000, cor = FALSE, corFunc = c("spearman", "pearson"),
  combp = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{dat.m}{450k methylation microarray matrix via preprocessing; m x n matrix}

\item{design}{design matrix; [1, covariate of interested, paired info]}

\item{sv.m}{surrogate variables matrix}

\item{chr}{chromosome vector for probes}

\item{pos}{position vector for probes}

\item{cluster}{The clusters of locations that are to be analyzed together. via correlation
perspevtive, we can use agglomerate cluster index to bump analysis. If cluster is not
available, clusterMaker can be used to cluster nearby locations}

\item{coef}{Integer. covariate of interest's column}

\item{cutoff}{numeric value. Value of estimates coefficient of covariate of interested above cutoff
or below the negative of cutoff will be used as candidate bump regions.}

\item{pvalue}{numeric value. cutoff of pvalue for candidate regions selection}

\item{maxGap}{if cluster is not availabel. maxGap is used by clusterMaker to define cluster}

\item{minDist}{if clusters are build, the mininal distance between clusters is setted as constraint}

\item{nullMethod}{Method for generating null candidate regions. If ncol(design) > 2. bootstrap method is recommanded}

\item{robust}{logic, use robust linear regression or not; Default FALSE}

\item{smooth}{logic. If TRUE then the standard error or correlation of point-wise estimatrs will be used as weigths
in the \code{loessByCluster}}

\item{smoother}{smooth function to estimate genome profile.}

\item{B}{integer; Denoting the number of resamples to computr null distribution. Default = 0
B is used to generate permutation matrix,
which describes permutations to generate null distribution}

\item{cor}{logical; If TRUE then position correlation matrix is considered as weighted matrix}

\item{corFunc}{Optional; "spearman"(Default) and "pearson"}

\item{combp}{logical; If TRUE then slk p correction will be applied}

\item{verbose}{logical. Optional printing progress message or not}

\item{...}{}
}
\value{
bumps object
}
\description{
bump hunting algorithm for 450k methylation microarray
}
\details{
bumphuntingEngine
}
\examples{
# cluster input format: cg12045430 cg20826792 cg00381604 cg20253340
}

