% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/bumphunter.R
\docType{methods}
\name{get.bumps}
\alias{get.bumps}
\alias{get.bumps,bumps-method}
\title{get.bumps method}
\usage{
get.bumps(object, ...)

\S4method{get.bumps}{bumps}(object, ...)
}
\arguments{
\item{object}{methylation expression matrix object}

\item{...}{additional parameters}
}
\value{
bumps object
}
\description{
get.bumps method

get.bumps
}

