% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{List2Index}
\alias{List2Index}
\title{List2Index}
\usage{
List2Index(cluster, chr, pos, names)
}
\arguments{
\item{cluster}{vector of probes clusters By distance(maxGap) constraints}

\item{chr}{Chromosome vector}

\item{pos}{position numeric vector}

\item{names}{probe names vector ; used in function \link{clusterMaker}}
}
\value{
clusterIndex
}
\description{
convert cluster list to Index with increased number
}

